Improved emoji regex.
The previous regex was not removing emojis that were a traditional pictograph followed by an emoji specifier codepoint. The new regex handles this case, an example would be the ⚔️ which is really two codepoints, a black and white crossed swords, followed by a special codepoint that says "draw this as an emoji instead".