Renamed PeopleCard to CharacterCard, set minHeight
There was no minHeight set on the CharacterCard so when discussing AI created characters that don't have portraits, the card ended up shrinking down. Technically a regression as I believe there used to be a minHeight awhile back.

Updated readme to reflect the renaming of PeopleCard to CharecterCard.

Updated props of CharacterCard to be `characters` instead of `people`.