Added ability to save and playback stories
This is still very rough, it is possible to break it in many ways, but the bones are there.

Stories are saved to S3, which s3 is configured through environment variables in the backend's .env. Feel free to use an s3 provider of your choosing.

Front end has a bunch of new components to handle this, including a component that fetches saved stories, changes to the initial input field so there is a place to enter the story id, and the ability to save stories on the story summary page.